Draft: fix clang compiler errors

See merge request git/vvenc!688

+++ source/App/vvencFFapp/EncApp.h
@@ -120,7 +120,7 @@ private:
   void closeFileIO();
 
   // statistics
-  void printRateSummary( int framesRcvd );
+  void printRateSummary( int64_t framesRcvd );
   void printChromaFormat();
 };
